Remove AstToActionSequence
