+++ configs/hearthstone/treegen_train.yaml
@@ -20,7 +20,7 @@ chars:
 samples:
   type: mlprogram.utils.data.get_samples
   dataset: "@/train_dataset"
-  to_action_sequence: "@/to_action_sequence"
+  parser: "@/parser"
 
 encoder:
   word_encoder:
@@ -73,7 +73,7 @@ transform:
           max_word_length: 128
       - - "transform_code"
         - type: mlprogram.utils.transform.action_sequence.TransformCode
-          to_action_sequence: "@/to_action_sequence"
+          parser: "@/parser"
       - - "transform_action_sequence"
         - type: mlprogram.utils.transform.treegen.TransformActionSequence
           action_sequence_encoder: "@/encoder.action_sequence_encoder"
