+++ mlprogram/actions/__init__.py
@@ -2,4 +2,3 @@ from mlprogram.actions.action import NodeConstraint, NodeType  # noqa
 from mlprogram.actions.action import ExpandTreeRule, CloseVariadicFieldRule, Rule  # noqa
 from mlprogram.actions.action import ApplyRule, GenerateToken, Action  # noqa
 from mlprogram.actions.action_sequence import InvalidActionException, ActionSequence  # noqa
-from mlprogram.actions.ast_to_action_sequence import AstToActionSequence  # noqa
