+++ test/languages/csg/test_functions.py
@@ -1,5 +1,4 @@
 import unittest
-from mlprogram.actions import AstToActionSequence
 from mlprogram.languages.csg import IsSubtype
 from mlprogram.languages.csg import Parser
 from mlprogram.languages.csg import get_samples, Dataset
@@ -19,10 +18,7 @@ class TestGetSamples(unittest.TestCase):
     def test_not_reference(self):
         dataset = Dataset(1, 1, 1, 1, 45)
 
-        def to_action_sequence(code):
-            return AstToActionSequence()(Parser().parse(code))
-
-        samples = get_samples(dataset, to_action_sequence)
+        samples = get_samples(dataset, Parser())
         self.assertEqual(7, len(samples.rules))
         self.assertEqual(12, len(samples.node_types))
         self.assertEqual(9, len(samples.tokens))
@@ -30,10 +26,7 @@ class TestGetSamples(unittest.TestCase):
     def test_reference(self):
         dataset = Dataset(1, 1, 1, 1, 45, reference=True)
 
-        def to_action_sequence(code):
-            return AstToActionSequence()(Parser().parse(code))
-
-        samples = get_samples(dataset, to_action_sequence)
+        samples = get_samples(dataset, Parser())
         self.assertEqual(7, len(samples.rules))
         self.assertEqual(12, len(samples.node_types))
         self.assertEqual(9, len(samples.tokens))
