+++ test/utils/data/test_functions.py
@@ -1,11 +1,9 @@
 import torch
 import unittest
 import numpy as np
-import ast
 from typing import List
-from mlprogram.actions import AstToActionSequence
 from mlprogram.languages import Token
-from mlprogram.languages.python.python_ast_to_ast import to_ast
+from mlprogram.languages.python import Parser
 from mlprogram.utils.data \
     import ListDataset, get_samples, get_words, \
     get_characters, Collate, CollateOptions
@@ -13,11 +11,6 @@ from mlprogram.utils.data \
 
 def tokenize(str: str) -> List[Token]:
     return list(map(lambda x: Token(None, x, x), str.split(" ")))
-
-
-def to_action_sequence(code: str):
-    return AstToActionSequence()(
-        to_ast(ast.parse(code).body[0], lambda x: [x]))
 
 
 class TestGetWords(unittest.TestCase):
@@ -47,7 +40,7 @@ class TestGetSamples(unittest.TestCase):
         entries = [{"input": ["foo bar"], "ground_truth": ["y = x + 1"]},
                    {"input": ["test foo"], "ground_truth": ["f(x)"]}]
         dataset = ListDataset(entries)
-        d = get_samples(dataset, to_action_sequence)
+        d = get_samples(dataset, Parser(lambda x: [x]))
         self.assertEqual([
             ("str", "y"),
             ("str", "x"),
