+++ test/utils/transform/test_action_sequence.py
@@ -4,7 +4,7 @@ import numpy as np
 from mlprogram.utils.data import ListDataset, get_samples
 from mlprogram.languages import Node, Leaf, Field
 from mlprogram.languages import Token
-from mlprogram.actions import AstToActionSequence
+from mlprogram.languages import Parser
 from mlprogram.encoders import ActionSequenceEncoder
 from mlprogram.utils.transform.action_sequence \
     import TransformCode, TransformGroundTruth, \
@@ -15,27 +15,30 @@ def tokenize(query: str):
     return list(map(lambda x: Token(None, x, x), query.split(" ")))
 
 
-def to_action_sequence(code: str):
-    ast = Node("Assign",
-               [Field("name", "Name",
-                      Node("Name", [Field("id", "str", [Leaf("str", "x")])])),
-                Field("value", "expr",
-                      Node("Op", [
-                           Field("op", "str", [Leaf("str", "+")]),
-                           Field("arg0", "expr",
-                                 Node("Name", [Field("id", "str",
-                                                     [Leaf("str", "y")])])),
-                           Field("arg1", "expr",
-                                 Node("Number", [Field("value", "number",
-                                                       [Leaf("number", "1")])
-                                                 ]))]
-                           ))])
-    return AstToActionSequence()(ast)
+class MockParser(Parser[str]):
+    def parse(self, code: str):
+        ast = Node("Assign",
+                   [Field("name", "Name",
+                          Node("Name", [Field("id", "str",
+                                              [Leaf("str", "x")])])),
+                    Field("value", "expr",
+                          Node("Op", [
+                              Field("op", "str", [Leaf("str", "+")]),
+                              Field("arg0", "expr",
+                                    Node("Name", [Field("id", "str",
+                                                        [Leaf("str", "y")])])),
+                              Field("arg1", "expr",
+                                    Node("Number", [
+                                        Field("value", "number",
+                                              [Leaf("number", "1")])
+                                    ]))]
+                               ))])
+        return ast
 
 
 class TestTransformCode(unittest.TestCase):
     def test_simple_case(self):
-        transform = TransformCode(to_action_sequence)
+        transform = TransformCode(MockParser())
         action_sequence = \
             transform({"ground_truth": "y = x + 1"})["action_sequence"]
         self.assertEqual(None, action_sequence.head)
@@ -45,10 +48,10 @@ class TestTransformGroundTruth(unittest.TestCase):
     def test_simple_case(self):
         entries = [{"input": "foo bar", "ground_truth": "y = x + 1"}]
         dataset = ListDataset(entries)
-        d = get_samples(dataset, to_action_sequence)
+        d = get_samples(dataset, MockParser())
         aencoder = ActionSequenceEncoder(d, 0)
         input = \
-            TransformCode(to_action_sequence)({"ground_truth": "y = x + 1"})
+            TransformCode(MockParser())({"ground_truth": "y = x + 1"})
         transform = TransformGroundTruth(aencoder)
         input["reference"] = [
             Token(None, "foo", "foo"), Token(None, "bar", "bar")]
@@ -67,10 +70,10 @@ class TestTransformGroundTruth(unittest.TestCase):
     def test_impossible_case(self):
         entries = [{"input": "foo bar", "ground_truth": "y = x + 1"}]
         dataset = ListDataset(entries)
-        d = get_samples(dataset, to_action_sequence)
+        d = get_samples(dataset, MockParser())
         d.tokens = [("", "y"), ("", "1")]
         aencoder = ActionSequenceEncoder(d, 0)
-        action_sequence = TransformCode(to_action_sequence)({
+        action_sequence = TransformCode(MockParser())({
             "ground_truth": "y = x + 1"
         })["action_sequence"]
         transform = TransformGroundTruth(aencoder)
@@ -85,10 +88,10 @@ class TestTransformActionSequenceForRnnDecoder(unittest.TestCase):
     def test_simple_case(self):
         entries = [{"input": "foo bar", "ground_truth": "y = x + 1"}]
         dataset = ListDataset(entries)
-        d = get_samples(dataset, to_action_sequence)
+        d = get_samples(dataset, MockParser())
         aencoder = ActionSequenceEncoder(d, 0)
         transform = TransformActionSequenceForRnnDecoder(aencoder)
-        action_sequence = TransformCode(to_action_sequence)({
+        action_sequence = TransformCode(MockParser())({
             "ground_truth": "y = x + 1"
         })["action_sequence"]
         result = transform({
@@ -109,9 +112,9 @@ class TestTransformActionSequenceForRnnDecoder(unittest.TestCase):
     def test_eval(self):
         entries = [{"input": "foo bar", "ground_truth": "y = x + 1"}]
         dataset = ListDataset(entries)
-        d = get_samples(dataset, to_action_sequence)
+        d = get_samples(dataset, MockParser())
         aencoder = ActionSequenceEncoder(d, 0)
-        action_sequence = TransformCode(to_action_sequence)({
+        action_sequence = TransformCode(MockParser())({
             "ground_truth": "y = x + 1"
         })["action_sequence"]
         transform = TransformActionSequenceForRnnDecoder(aencoder, train=False)
@@ -129,11 +132,11 @@ class TestTransformActionSequenceForRnnDecoder(unittest.TestCase):
     def test_impossible_case(self):
         entries = [{"input": "foo bar", "ground_truth": "y = x + 1"}]
         dataset = ListDataset(entries)
-        d = get_samples(dataset, to_action_sequence)
+        d = get_samples(dataset, MockParser())
         d.tokens = [("", "y"), ("", "1")]
         aencoder = ActionSequenceEncoder(d, 0)
         transform = TransformActionSequenceForRnnDecoder(aencoder)
-        action_sequence = TransformCode(to_action_sequence)({
+        action_sequence = TransformCode(MockParser())({
             "ground_truth": "y = x + 1"
         })["action_sequence"]
         result = transform({
